랜딩 구단메인 프로토콜 추가

+++ Assets/Scripts/Common/Models/QuizRespInfo.cs
@@ -146,9 +146,9 @@ public class QuizRespInfo {
 		}
 	}
 
-	int _addRewardRate;
+	float _addRewardRate;
 
-	public int addRewardRate {
+	public float addRewardRate {
 		get {
 			return _addRewardRate;
 		}
